refactor: 이벤트 핸들러 view로 이동
